+++ ui/src/app/machines/views/MachineDetails/MachineSummary/SummaryNotifications/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SummaryNotifications";
